app/services/auth_service.py
"""
Authentication service for user management and login.

This module supports two authentication types:
- Email/password authentication (custom backend)
- Google/Facebook login via Firebase OAuth (using Firebase Admin SDK)

Also includes functionality for:
- JWT access token generation and validation
- Password hashing and verification
- Auto-registration of users via Firebase OAuth login
"""

from datetime import datetime, timedelta
from typing import Optional

from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy.orm import Session
from fastapi import HTTPException, status

from app.db.models.user import User
from app.schemas.user import UserCreate

import firebase_admin
from firebase_admin import credentials, auth as firebase_auth

# Password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# JWT Configuration
SECRET_KEY = "your_secret_key_here"  # Replace with secure .env variable
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60

# Initialize Firebase Admin SDK (replace with your actual JSON credentials)
cred = credentials.Certificate("firebase_credentials.json")
firebase_admin.initialize_app(cred)



def get_password_hash(password: str) -> str:
    """
    Hash a plain-text password using bcrypt.

    Args:
        password (str): The user's plain password.

    Returns:
        str: A securely hashed password string.
    """
    
    pass

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Verify a plain password against its hashed version.

    Args:
        plain_password (str): Raw password entered by the user.
        hashed_password (str): Stored hashed password from the database.

    Returns:
        bool: True if match, False otherwise.
    """
    
    pass

def register_user(db: Session, user_data: UserCreate) -> User:
    """
    Register a new user using email and password.

    Args:
        db (Session): Active SQLAlchemy database session.
        user_data (UserCreate): User registration details.

    Returns:
        User: The newly created user record.

    Raises:
        HTTPException: If the email is already in use.
    """
    
    pass

def authenticate_user(db: Session, email: str, password: str) -> Optional[User]:
    """
    Authenticate a user with email and password.

    Args:
        db (Session): Active database session.
        email (str): Email address.
        password (str): Raw password.

    Returns:
        Optional[User]: User object if credentials are valid, else None.
    """
    
    pass

# --------------------------- JWT SESSION MANAGEMENT --------------------------- #

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    Generate a JWT token with expiration.

    Args:
        data (dict): Data to encode in the token (e.g., {"sub": user_id}).
        expires_delta (Optional[timedelta]): Custom expiration duration.

    Returns:
        str: Encoded JWT token string.
    """
    
    pass

def decode_token(token: str) -> Optional[dict]:
    """
    Decode a JWT token and validate its signature and expiration.

    Args:
        token (str): JWT token string.

    Returns:
        Optional[dict]: Decoded payload if valid, else None.
    """
    
    pass

# --------------------------- FIREBASE OAUTH AUTH --------------------------- #

def verify_oauth_token(id_token: str) -> dict:
    """
    Verify a Firebase ID token for OAuth (Google/Facebook).

    Args:
        id_token (str): Firebase ID token received from frontend.

    Returns:
        dict: Decoded token payload including uid, email, name, etc.

    Raises:
        HTTPException: If the token is invalid.
    """
    
    pass

def login_or_register_oauth(db: Session, firebase_data: dict) -> User:
    """
    Log in or auto-register a user using Firebase token data.

    Args:
        db (Session): SQLAlchemy DB session.
        firebase_data (dict): Firebase token payload.

    Returns:
        User: The logged-in or newly created user.
    """
    
    pass
