pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant-reservation-system"
version = "0.1.0"
description = "Full-featured Restaurant Reservation & Delivery System with FastAPI, Firebase, Email, and Admin support."
authors = [
    { name = "Youssef Osama", email = "osama74454@gmail.com" },
    { name = "Your Name", email = "you@example.com" },
    { name = "Your Name", email = "you@example.com" },
    { name = "Your Name", email = "you@example.com" },
    { name = "Your Name", email = "you@example.com" },
    { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.29",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "PyJWT==2.8.0",
    "firebase-admin==6.5.0",
    "google-auth==2.29.0",
    "python-dotenv==1.0.1",
    "email-validator==2.1.1",
    "jinja2==3.1.3",
    "fastapi-utils==0.2.1",
    "httpx==0.27.0",
    "requests==2.31.0",
    "loguru==0.7.2",
    "python-slugify==8.0.4",
    "ipython==8.24.0",
    "markdown==3.5.2",
    "itsdangerous==2.1.2"
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "isort==5.13.2",
    "autopep8==2.3.2",
    "pylint==3.3.6",
    "flake8==7.1.2",
    "mypy==1.15.0",
    "bandit==1.7.7",
    "radon==6.0.1",
    "pytest==8.1.1",
    "pytest-cov==5.0.0",
    "httpx[http2]==0.27.0"
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov"
testpaths = ["tests"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 88
exclude = [".git", "__pycache__", "build", "dist"]

[tool.bandit]
skips = ["B101"]  

